backend/requirements.txt
fastapi
uvicorn
chromadb
httpx
python-dotenv
